+++ src/random.h
@@ -126,7 +126,7 @@ private:
     int bytebuf_size;
 
     uint64_t bitbuf;
-    int bitbuf_size;
+    uint64_t bitbuf_size;
 
     void RandomSeed();
 
@@ -169,7 +169,7 @@ public:
     }
 
     /** Generate a random (bits)-bit integer. */
-    uint64_t randbits(int bits) noexcept
+    uint64_t randbits(uint64_t bits) noexcept
     {
         if (bits == 0) {
             return 0;
@@ -191,7 +191,7 @@ public:
     {
         assert(range);
         --range;
-        int bits = CountBits(range);
+        uint64_t bits = CountBits(range);
         while (true) {
             uint64_t ret = randbits(bits);
             if (ret <= range) return ret;
@@ -202,7 +202,7 @@ public:
     std::vector<unsigned char> randbytes(size_t len);
 
     /** Generate a random 32-bit integer. */
-    uint32_t rand32() noexcept { return randbits(32); }
+    uint32_t rand32() noexcept { return static_cast<uint32_t>(randbits(32)); }
 
     /** generate a random uint256. */
     uint256 rand256() noexcept;
